Migrate from web3 to ethers
Implement sendTx, restore and connect
