+++ index.ts
@@ -1,0 +1,5 @@
+import WalletProvider from './WalletProvider'
+
+export { WalletContext } from './WalletContext'
+export { shortenAddress, isValidAddress } from './utils'
+export default WalletProvider
