+++ networks.ts
@@ -317,7 +317,7 @@ const networksRaw = [
     currency_symbol: 'DEV',
     currency_decimals: 18,
     is_testnet: true
-  },
+  }
 ]
 
 const networks = networksRaw.map(item => ({
@@ -342,6 +342,11 @@ const networks = networksRaw.map(item => ({
     ]
   }
 }))
+
+export const rpcMapping = networksRaw.reduce((mapper, network) => {
+  mapper[network.chain_id] = network.rpc_url
+  return mapper
+}, {})
 
 export const getNetworkById = (chainId: string | number) => {
   const network = networks.find(net => net.chain_id === chainId)
